core/requirements.txt
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.8.0rc2
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-api-response==0.1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0
zope.interface==6.1
